pyproject.toml
[project]
name = "finops-autopilot"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "boto3",
    "typer[all]",
    "rich",
    "pyyaml",
    "jsonschema",
    "duckdb",
    "jinja2",
    "kubernetes",
    "requests",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "moto[all]",
    "ruff",
    "mypy",
]

[project.scripts]
finops = "finops.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
